requirements.txt
pandas==1.3.5
googlemaps==4.5.3
Unidecode==1.3.2
numpy==1.22.1
python_dotenv==0.19.2